Move buildinfo pkg out of internal.
Intended to share this module with verification server.
